refactor: the decoration passes move out of CityGenerator (#11)
gen/Decorations holds what is left lying about: rubble, ruined upper
floors, undergrowth and park planting - plus the height interpolation
only rubble used, and the positional roll all four share.

The three noise fields moved with them. Each is a pure function of the
world seed, describes the whole dimension, and was read by nothing else,
so they belong to the passes rather than to the generator. Their seed
offsets are positional and stay so - rubble 0, leaves 1, ruins 2; the
fourth, the floating profile's underside, stays with the terrain code and
keeps offset 3.

This is what extracting GroundCover first was for: these four were the
only remaining readers of the leaf and rubble tables, so they now move
without widening anything.

Pure code motion, verified token-for-token against the originals modulo
the added parameter and the renames: 222, 267, 389, 440 and 387 tokens.

CityGenerator 2201 -> 1891 lines.

All seven digest suites unchanged.
